refactor: Clean Architecture 구조로 인증 시스템 재구성
- 중복된 AuthSchemas.ts 파일 제거
- 모듈화된 검증 필드로 분리 (EmailField, PasswordField, NameField, UserTypeField)
- Clean Architecture에 맞춰 application/dto, application/usecase 구조로 정리
- 모든 import 경로 업데이트
- 테스트 파일 import 경로 수정
- 재사용 가능한 validation 컴포넌트 구조 구현
