+++ app/api/auth/[...nextauth]/application/dto/AuthError.ts
