+++ app/api/auth/[...nextauth]/application/dto/ValidationError.ts
