+++ app/api/auth/[...nextauth]/application/dto/index.ts
