+++ app/api/auth/[...nextauth]/application/usecase/index.ts
