--- app/api/auth/[...nextauth]/validation/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthSchemas';
