+++ app/api/auth/__tests__/infrastructure/next-auth/AuthConfig.simple.test.ts
@@ -1,7 +1,7 @@
 import { authOptions } from "@/app/api/infrastructure/next-auth/AuthConfig";
 
 // VerifyCredentialsUseCase와 SbUserRepository 모킹
-jest.mock('../../../[...nextauth]/usecase/VerifyCredentialsUseCase');
+jest.mock('../../../[...nextauth]/application/usecase/VerifyCredentialsUseCase');
 jest.mock('../../../../infrastructure/repositories/SbUserRepository');
 
 describe('NextAuth Configuration (Simple)', () => {
