+++ app/api/auth/__tests__/infrastructure/repositories/SupabaseUserRepository.test.ts
@@ -2,7 +2,7 @@ import '../../__mocks__/supabase.mock';
 import { SupabaseUserRepository } from '../../../../infrastructure/repositories/SbUserRepository';
 import { mockSupabaseClient } from '../../__mocks__/supabase.mock';
 import { User } from '../../../../domain/entities/User';
-import { SignupRequest } from '../../signup/dto/SignupRequest';
+import { SignupRequest } from '../../signup/application/dto/SignupRequest';
 
 describe('SupabaseUserRepository', () => {
   let repository: SupabaseUserRepository;
