+++ app/api/auth/__tests__/integration/auth.integration.test.ts
@@ -1,7 +1,7 @@
 import '../__mocks__/supabase.mock';
 import '../__mocks__/bcrypt.mock';
-import { RegisterUseCase } from '../../signup/usecase/RegisterUseCase';
-import { LoginUseCase } from '../../[...nextauth]/usecase/LoginUseCase';
+import { RegisterUseCase } from '../../signup/application/usecase/RegisterUseCase';
+import { LoginUseCase } from '../../[...nextauth]/application/usecase/LoginUseCase';
 import { SupabaseUserRepository } from '../../../infrastructure/repositories/SbUserRepository';
 import { mockSupabaseClient } from '../__mocks__/supabase.mock';
 import { mockBcrypt } from '../__mocks__/bcrypt.mock';
