+++ app/api/auth/signup/application/dto/AuthError.ts
