+++ app/api/auth/signup/application/dto/WeakPasswordError.ts
